vz-example-viewer: fix int feature list type
Summary:
Caught during a Clutz upgrade. From context, it looks like `ints` should
be an `Int64List`, as is already the case in `parseFeature` below.

Test Plan:
Fingers crossed; I’m not familiar with this part of the codebase.

wchargin-branch: vzev-ints-list-type

+++ tensorboard/components/vz_example_viewer/vz-example-viewer.ts
@@ -922,7 +922,7 @@ Polymer({
     let jsonFeat: any;
     if (this.newFeatureType === INT_FEATURE_NAME) {
       const valueList: number[] = [];
-      const ints = new FloatList();
+      const ints = new Int64List();
       ints.setValueList(valueList);
       feat.setInt64List(ints);
       jsonFeat = {int64List: {value: valueList}};
